Moved Connection to terminal API
