+++ readline/src/main/java/org/aesh/readline/tty/terminal/TerminalConnection.java
@@ -25,7 +25,7 @@ import org.aesh.terminal.Attributes;
 import org.aesh.terminal.Terminal;
 import org.aesh.readline.terminal.TerminalBuilder;
 import org.aesh.terminal.tty.Capability;
-import org.aesh.readline.tty.Connection;
+import org.aesh.terminal.Connection;
 import org.aesh.util.LoggerUtil;
 
 import java.io.IOException;
