+++ terminal-api/src/main/java/org/aesh/terminal/Connection.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.aesh.readline.tty;
+package org.aesh.terminal;
 
 import org.aesh.terminal.tty.Size;
 import org.aesh.terminal.tty.Capability;
